# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "webstie titlte",
        "description": "API documention",
        "termsOfService": "https://example.com",
        "contact": {
            "name": "mahdi",
            "email": "mahdishoorabi@gmail.com",
            "url": "https://example.com"
        },
        "license": {
            "name": "License Name",
            "url": "https://example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/routines": {
            "get": {
                "tags": [
                    "routines"
                ],
                "summary": "Get list of routines",
                "description": "Returns list of routines",
                "operationId": "6f4b99b6a840eac418ea0f6f4614f133",
                "responses": {
                    "200": {
                        "description": "A list of routines"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "routines"
                ],
                "summary": "Create a new routine",
                "description": "Stores a new routine",
                "operationId": "443061fe2238d4ba22520868adb2e505",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Routine created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/api/v1/routines/{id}": {
            "put": {
                "tags": [
                    "routines"
                ],
                "summary": "Update an existing routine",
                "description": "Updates a routine by ID",
                "operationId": "031346e7cbfbaecef010e74b406a6398",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the routine to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Routine updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "routines"
                ],
                "summary": "Delete a routine",
                "description": "Deletes a routine by ID",
                "operationId": "1abcee3f6137df1e8c9282abefe8d5e5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the routine to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Routine deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/v1/sessions": {
            "get": {
                "tags": [
                    "sessions"
                ],
                "summary": "Get list of sessions",
                "description": "Returns list of sessions",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "sessions"
                ],
                "summary": "Create new session",
                "description": "Stores a new session",
                "operationId": "store",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Session created"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/api/v1/sessions/{id}": {
            "put": {
                "tags": [
                    "sessions"
                ],
                "summary": "Update an existing session",
                "description": "Updates a session by ID",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Session ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Session updated"
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "sessions"
                ],
                "summary": "Delete a session",
                "description": "Deletes a session by ID",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Session ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Session deleted"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
...
}